More New Boss Cards + Card Adjustments
Added more boss cards, in order to split up the effects of the boss. Gave some text adjustments to some cards. The Sneak Peek pack was also changed to be more general, as I am going to be adding more cards to it in the future that may not be entirely related to the anime.